fix(native): move polyfills to canvas, enable tree-shaking
